[Export] Refactor export dialog: separate ui from business logic
