+++ src/export/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_library(mediaelch_export OBJECT ExportTemplate.cpp ExportTemplateLoader.cpp
-                                    TableWriter.cpp)
+                                    MediaExport.cpp TableWriter.cpp)
 
 target_link_libraries(
   mediaelch_export
